Add solution classes for longest increasing subsequence problems
Added two new solution classes, one for the problem of finding the length of the longest increasing subsequence in an integer array, and the other for finding the number of longest increasing subsequences. Corresponding tests are added for both solutions. The implemented algorithms are based on dynamic programming, where each array element scans all the elements on its left to determine the max length and count.