Add solution for minimum index sum of two lists problem
Implemented a new algorithm on leetcode called minimum index sum of two lists. The solution involves using a HashMap to store elements from the first list and their indices. Then, in a second iteration, the minimum index sum is calculated. If a lower index sum is found, the list of results is cleared and the current restaurant is added. If the index sum is equal to the minimum, the restaurant is added to the list of results. An empty test class is also added and will be filled later.